+++ skrub/_wrap_transformer.py
@@ -1,5 +1,5 @@
-from ._apply_to_cols import ApplyToCols
-from ._apply_to_frame import ApplyToFrame
+from ._apply_on_each_col import ApplyToEachCol
+from ._apply_sub_frame import ApplyToSubFrame
 from ._single_column_transformer import is_single_column_transformer
 from .selectors import make_selector
 
@@ -15,19 +15,19 @@ def wrap_transformer(
     n_jobs=None,
     columnwise="auto",
 ):
-    """Create a ``ApplyToCols`` or a ``ApplyToFrame`` transformer.
+    """Create a ``ApplyToEachCol`` or a ``ApplyToSubFrame`` transformer.
 
     The ``transformer`` is wrapped in a transformer that will apply it to part
     of the input dataframe.
 
     By default, if ``transformer`` is a single-column transformer
     (has a ``__single_column_transformer__`` attribute), it is wrapped in a
-    ``ApplyToCols`` instance. Otherwise it is wrapped in a ``ApplyToFrame``
+    ``ApplyToEachCol`` instance. Otherwise it is wrapped in a ``ApplyToSubFrame``
     instance.
 
     This default choice can be overridden by passing ``columnwise=True`` to
-    force the use of ``ApplyToCols`` or ``columnwise=False`` to force the use
-    of ``ApplyToFrame``.
+    force the use of ``ApplyToEachCol`` or ``columnwise=False`` to force the use
+    of ``ApplyToSubFrame``.
 
     Parameters
     ----------
@@ -39,26 +39,26 @@ def wrap_transformer(
 
     allow_reject : bool, default=False
         Whether to allow column rejections. Only used when the result is an
-        instance of ``ApplyToCols``, see this class' docstring for details.
+        instance of ``ApplyToEachCol``, see this class' docstring for details.
 
     keep_original : bool, default=False
         Whether to retain the original columns in transformed output. See the
-        documentation of ``ApplyToCols`` or ``ApplyToFrame`` for details.
+        documentation of ``ApplyToEachCol`` or ``ApplyToSubFrame`` for details.
 
     rename_columns : str, default='{}'
         Format string applied to output column names. See the documentation of
-        ``ApplyToCols`` or ``ApplyToFrame`` for details.
+        ``ApplyToEachCol`` or ``ApplyToSubFrame`` for details.
 
     n_jobs : int, default=None
         Number of jobs to run in parallel. Only used when the result is an
-        instance of ``ApplyToCols``, see this class' docstring for details.
+        instance of ``ApplyToEachCol``, see this class' docstring for details.
 
     columnwise : 'auto' or bool, default='auto'
-        Whether to create a ``ApplyToCols`` or ``ApplyToFrame`` instance. By
-        default, ``ApplyToCols`` is used if ``transformer`` has a
-        ``__single_column_transformer__`` attribute and ``ApplyToFrame``
-        otherwise. Pass ``columnwise=True`` to force using ``ApplyToCols`` and
-        ``columnwise=False`` to force using ``ApplyToFrame``. Note that forcing
+        Whether to create a ``ApplyToEachCol`` or ``ApplyToSubFrame`` instance. By
+        default, ``ApplyToEachCol`` is used if ``transformer`` has a
+        ``__single_column_transformer__`` attribute and ``ApplyToSubFrame``
+        otherwise. Pass ``columnwise=True`` to force using ``ApplyToEachCol`` and
+        ``columnwise=False`` to force using ``ApplyToSubFrame``. Note that forcing
         ``columnwise=False`` for a single-column transformer will most likely
         cause an error during ``fit``, and forcing ``columnwise=True`` for a
         regular transformer is only appropriate if the transformer can be
@@ -68,7 +68,7 @@ def wrap_transformer(
     Returns
     -------
     Wrapped transformer
-        A ``ApplyToCols`` or ``ApplyToFrame`` instance initialized with the
+        A ``ApplyToEachCol`` or ``ApplyToSubFrame`` instance initialized with the
         input ``transformer``.
 
     Examples
@@ -79,11 +79,11 @@ def wrap_transformer(
     >>> from sklearn.preprocessing import OrdinalEncoder
 
     >>> wrap_transformer(ToDatetime(), s.all())
-    ApplyToCols(transformer=ToDatetime())
+    ApplyToEachCol(transformer=ToDatetime())
     >>> wrap_transformer(OrdinalEncoder(), s.string())
-    ApplyToFrame(cols=string(), transformer=OrdinalEncoder())
+    ApplyToSubFrame(cols=string(), transformer=OrdinalEncoder())
     >>> wrap_transformer(OrdinalEncoder(), s.string(), columnwise=True, n_jobs=4)
-    ApplyToCols(cols=string(), n_jobs=4, transformer=OrdinalEncoder())
+    ApplyToEachCol(cols=string(), n_jobs=4, transformer=OrdinalEncoder())
     """
     selector = make_selector(selector)
 
@@ -91,7 +91,7 @@ def wrap_transformer(
         columnwise = is_single_column_transformer(transformer)
 
     if columnwise:
-        return ApplyToCols(
+        return ApplyToEachCol(
             transformer,
             cols=selector,
             allow_reject=allow_reject,
@@ -99,7 +99,7 @@ def wrap_transformer(
             rename_columns=rename_columns,
             n_jobs=n_jobs,
         )
-    return ApplyToFrame(
+    return ApplyToSubFrame(
         transformer,
         cols=selector,
         keep_original=keep_original,