+++ CHANGES.rst
@@ -22,6 +22,9 @@ New Features
 
 Changes
 -------
+- ``ApplyToCols`` and ``ApplyToFrame`` have been renamed :class:`ApplyToEachCol`
+  and :class:`ApplyToSubFrame` respectively.
+  :pr:`1913` by :user:`Riccardo Cappuzzo <rcap107>`.
 
 Bug Fixes
 --------